Move d3-context-menu to collected D3 folder
